requirements_test.txt
apache-airflow-providers-celery
numpy==1.26.4
pandas==1.5.3
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
flake8==7.1.1